% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afpca-package.R
\docType{package}
\name{afpca-package}
\alias{afpca}
\alias{afpca-package}
\title{afpca: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/angelgar/afpca}
  \item Report bugs at \url{https://github.com/angelgar/afpca/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
