% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.searchOS.R
\name{arg.searchOS}
\alias{arg.searchOS}
\title{Generate Spline Basis to estimate FPC}
\usage{
arg.searchOS(string, arg.name)
}
\description{
This is to generate default knots
This code belongs to AdaptFitOS
}
\keyword{internal}
