% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspmFormReadOS.R
\name{aspmFormReadOS}
\alias{aspmFormReadOS}
\title{Generate Spline Basis to estimate FPC}
\usage{
aspmFormReadOS(form, omit.missing = FALSE, constrasts = NULL)
}
\description{
This is to generate spline basis
This code belongs to AdaptFitOS
}
\keyword{internal}
