% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspmKrigeReadOS.R
\name{aspmKrigeReadOS}
\alias{aspmKrigeReadOS}
\title{Generate Spline Basis to estimate FPC}
\usage{
aspmKrigeReadOS(term)
}
\description{
This is to generate default knots
This code belongs to AdaptFitOS
}
\keyword{internal}
