% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspmPenReadOS.R
\name{aspmPenReadOS}
\alias{aspmPenReadOS}
\title{Generate Spline Basis to estimate FPC}
\usage{
aspmPenReadOS(term)
}
\description{
This is to generate default knots
This code belongs to AdaptFitOS
}
\keyword{internal}
