% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_cols.R
\name{center_cols}
\alias{center_cols}
\title{Column-centre a matrix  (replicates the centering step in spmDesignOS)}
\usage{
center_cols(M)
}
\description{
Column-centre a matrix  (replicates the centering step in spmDesignOS)
}
\keyword{internal}
