% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.knots.R
\name{default.knots}
\alias{default.knots}
\title{Generate Spline Basis to estimate FPC}
\usage{
default.knots(x, num.knots, knotchoice = "quantiles")
}
\description{
This is to generate default knots
This code belongs to AdaptFitOS
}
\keyword{internal}
