% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_knots.R
\name{default_knots}
\alias{default_knots}
\title{Direct replacement for AdaptFitOS::default.knots() / SemiPar::default.knots()
Internal function to generate knots}
\usage{
default_knots(x, num.knots, knotchoice = "quantiles")
}
\arguments{
\item{x}{numeric vector (the predictor)}

\item{num.knots}{integer, number of knots; auto-computed when omitted}

\item{knotchoice}{"quantiles" (default) or "equidistant"}
}
\description{
Direct replacement for AdaptFitOS::default.knots() / SemiPar::default.knots()
Internal function to generate knots
}
\keyword{internal}
