% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basis.R
\name{generate_basis}
\alias{generate_basis}
\title{Title}
\usage{
generate_basis(
  data,
  poly.degree = 2,
  knots = NA,
  nbs = 40,
  basis = "trunc.poly"
)
}
\arguments{
\item{basis}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
