% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basis_helper.R
\name{generate_basis_helper}
\alias{generate_basis_helper}
\title{Generate Spline Basis to estimate FPC}
\usage{
generate_basis_helper(
  form,
  spar.method = "REML",
  contrasts = NULL,
  omit.missing = NULL,
  returnFit = FALSE,
  niter = 20,
  niter.var = 50,
  tol = 1e-06,
  tol.theta = 1e-06,
  control = NULL
)
}
\description{
This is to generate spline basis
This code belongs to AdaptFitOS
}
\keyword{internal}
