% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefficients.R
\name{get_coefficients}
\alias{get_coefficients}
\title{Title}
\usage{
get_coefficients(data, c_mat, Theta, Lambda, sigma, N.Unp.Basis, nbs)
}
\arguments{
\item{nbs}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
