% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fpcs.R
\name{get_fpcs}
\alias{get_fpcs}
\title{Title}
\usage{
get_fpcs(Theta, coef, nbs, n.comp, orthogonalize_fpcs, N.Unp.Basis)
}
\arguments{
\item{N.Unp.Basis}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
