% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lambda.R
\name{get_lambda}
\alias{get_lambda}
\title{Estimate the Adaptive Penalty Matrix (Lambda)}
\usage{
get_lambda(coef, n.comp, N.Unp.Basis)
}
\arguments{
\item{coef}{Coefficients}

\item{n.comp}{Number of FPCs to estimate}

\item{N.Unp.Basis}{Number of Unpenalized Basis}
}
\value{

}
\description{
Internal function to estimate the Adaptive penalty matrix
}
\keyword{internal}
