% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likelihood.R
\name{get_likelihood}
\alias{get_likelihood}
\title{Estimate the Full Likelihood of the FPCA Model (Internal)}
\usage{
get_likelihood(
  data,
  coef,
  Theta,
  c_mat,
  N.Unp.Basis,
  nbs,
  n.comp,
  lambda,
  sigma
)
}
\arguments{
\item{data}{Input Data i.e. Y}

\item{coef}{Estimated Spline Coefficients}

\item{Theta}{Matrix of Penalized Spline Basis}

\item{c_mat}{Matrix of Scores}

\item{N.Unp.Basis}{Matrix of Unpenalized Spline Basis}

\item{nbs}{Number of penalized spline basis}

\item{n.comp}{Number of FPC's that are being estimated}

\item{lambda}{Adaptive Smoothness Penalty Matrix}

\item{sigma}{Residual Variance}
}
\value{

}
\description{
Estimate the Full Likelihood of the FPCA Model (Internal)
}
