% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.R
\name{get_sigma}
\alias{get_sigma}
\title{Estimate residual variance}
\usage{
get_sigma(data, coef, Theta, c_mat, N.Unp.Basis, nbs)
}
\arguments{
\item{data}{input data}

\item{coef}{estimated spline coefficients}

\item{Theta}{spline basis matrix}

\item{c_mat}{scores matrix}

\item{N.Unp.Basis}{number of unpenalized spline basis}

\item{nbs}{total number of spline basis}
}
\value{

}
\description{
internal function to estimate resisual variance
}
\keyword{internal}
