% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_sqrt.R
\name{matrix_sqrt}
\alias{matrix_sqrt}
\title{Positive-definite matrix square root via SVD  (from SemiPar::matrix.sqrt)}
\usage{
matrix_sqrt(A)
}
\description{
Positive-definite matrix square root via SVD  (from SemiPar::matrix.sqrt)
}
\keyword{internal}
