% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthonormal_step.R
\name{orthonormal_step}
\alias{orthonormal_step}
\title{Title}
\usage{
orthonormal_step(c_mat, orthogonalize_scores, normalize.scores)
}
\arguments{
\item{normalize.scores}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
