% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os_desing_matrices.R
\name{os_design_matrices}
\alias{os_design_matrices}
\title{O'Sullivan spline decomposition into fixed (C) and random (Z) parts.
Mirrors Predict.matrix.lme() / smooth.construct.os.smooth.spec() from asp-internal.r.}
\usage{
os_design_matrices(x, nk, degree)
}
\arguments{
\item{x}{predictor vector (length n)}

\item{nk}{number of interior knots (only the COUNT matters; knots are forced equidistant)}

\item{degree}{length-2 vector c(p, q)  where p = B-spline degree, q = penalty order}
}
\description{
O'Sullivan spline decomposition into fixed (C) and random (Z) parts.
Mirrors Predict.matrix.lme() / smooth.construct.os.smooth.spec() from asp-internal.r.
}
\keyword{internal}
