% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.matrix.lme.R
\name{Predict.matrix.lme}
\alias{Predict.matrix.lme}
\title{Generate Spline Basis to estimate FPC}
\usage{
Predict.matrix.lme(object, data, drv = 0, center = T, ...)
}
\description{
This is to generate default knots
This code belongs to AdaptFitOS
}
\keyword{internal}
