% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.matrix.ospline.smooth.R
\name{Predict.matrix.ospline.smooth}
\alias{Predict.matrix.ospline.smooth}
\title{Generate Spline Basis to estimate FPC}
\usage{
Predict.matrix.ospline.smooth(object, data, drv = 0, ...)
}
\description{
This is to generate default knots
This code belongs to AdaptFitOS
}
\keyword{internal}
