% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_adaptive_functional_data.R
\name{simulate_adaptive_functional_data}
\alias{simulate_adaptive_functional_data}
\title{Simulate Adaptive Functional Data}
\usage{
simulate_adaptive_functional_data(
  n.tp = 200,
  N.subj = 20,
  var.scores = c(4, 1),
  noise.var = 0.2,
  seed.num = 1
)
}
\arguments{
\item{n.tp}{number of timepoints}

\item{N.subj}{number of subjects}

\item{var.scores}{variance of the scores (this is the value of the eigenvalues)}

\item{noise.var}{residual noise variance}

\item{seed.num}{seed number}
}
\value{

}
\description{
Simulate Adaptive Functional Data
}
