% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_design_matrices.R
\name{spline_design_matrices}
\alias{spline_design_matrices}
\title{Main replacement for asp2()'s design matrix extraction.
Internal function to generate knots}
\usage{
spline_design_matrices(x, knots, basis = "os", degree = 3)
}
\arguments{
\item{x}{numeric predictor vector (your X_temp)}

\item{knots}{numeric knot vector (from default_knots() or user-supplied)}

\item{basis}{"trunc.poly" | "tps" | "os"   (what you pass as 'basis' in f())}

\item{degree}{integer (or c(p,q) for "os")  (your poly.degree)}
}
\description{
Main replacement for asp2()'s design matrix extraction.
Internal function to generate knots
}
\keyword{internal}
