% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmArgReadOS.R
\name{spmArgReadOS}
\alias{spmArgReadOS}
\title{Generate Spline Basis to estimate FPC}
\usage{
spmArgReadOS(arg.assignment)
}
\description{
This is to generate default knots
This code belongs to AdaptFitOS
}
\keyword{internal}
