% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmDesignOS.R
\name{spmDesignOS}
\alias{spmDesignOS}
\title{Generate Spline Basis to estimate FPC}
\usage{
spmDesignOS(spm.info)
}
\description{
This is to generate spline basis
This code belongs to AdaptFitOS
}
\keyword{internal}
