% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R, R/test_helper.R
\name{test}
\alias{test}
\title{Test function}
\usage{
test(x = NA)

test(x = NA)
}
\arguments{
\item{x}{just the input}
}
\value{
print x

print x
}
\description{
Test function

Test function
}
\examples{

test("Hi")


test("Hi")

}
