% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afpca-package.R
\docType{package}
\name{afpca-package}
\alias{afpca}
\alias{afpca-package}
\title{afpca: Adaptive Functional Principal Component Analysis}
\description{
A method for integrating adaptive smoothness into functional principal component analysis
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/angelgar/afpca}
  \item Report bugs at \url{https://github.com/angelgar/afpca/issues}
}

}
\author{
\strong{Maintainer}: Angel Garcia de la Garza \email{ag3914@columbia.edu} (\href{https://orcid.org/0000-0001-6062-5190}{ORCID})

Authors:
\itemize{
  \item Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}

}
\keyword{internal}
