% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca.adapt.R
\name{fpca.adapt}
\alias{fpca.adapt}
\title{Adaptive Functional Principal Component Analysis}
\usage{
fpca.adapt(
  data,
  col_name = "name",
  basis = "trunc.poly",
  poly.degree = 2,
  knots = NA,
  nbs = 35,
  n.comp = 18,
  pve = 0.99,
  normalize.scores = TRUE,
  orthogonalize_scores = TRUE,
  orthogonalize_fpcs = TRUE,
  ntimes = 100
)
}
\arguments{
\item{data}{A data frame, matrix, or a tidyfun "tfd.reg" object of functional observations.}

\item{col_name}{Column name of the data. Required if Y is a tibble. The package expects the column to be a "dft" object from tidyfun}

\item{basis}{Type of spline basis. "os" for B-spline basis and "trunc.poly" for truncated polynomials (default)}

\item{poly.degree}{The degree of the the spline basis}

\item{knots}{Number of knots used to generate a spline basis to estimate the mean function and functional principal components}

\item{nbs}{Number of spline basis to generate mean function and functional principal components}

\item{n.comp}{prespecified number of principal components}

\item{pve}{proportion of variance explained; used to choose the number of functional principal components}

\item{normalize.scores}{normalize scores at each iteration (default is true, recommended)}

\item{orthogonalize_scores}{orthogonalize scores at each iteration (default is true, recommended)}

\item{orthogonalize_fpcs}{orthogonalize FPCs at each iteration (default is true, recommended)}

\item{ntimes}{Maximum number of iterations}
}
\value{
The function returns an object of class "afpca" that contains the following:

\item{Y}{The observed data. Pass either a matrix, a "dft" object from tidyverse,
or a tibble. If a tibble is passed, then "col_name" is required}
\item{Y_hat}{The reconstructed fitted values}
\item{mean}{a vector of the estimated mean function}
\item{fpcs}{a tp (number of timepoints) by npc (number of fpcs) matrix of the estimated eigenfunctions}
\item{scores}{an n (number of subjects) by npc (number of fpcs) matrix of estimated scores}
\item{eigenvalues}{estimated eigenvalues i.e. variances of fpc scores}
\item{Basis}{list containing penalized (Theta_beta) and penalized (Theta_b) spline basis}
\item{Theta}{concatenation of (Theta_beta) and penalized (Theta_b) spline basis}
\item{coef}{estimated coefficients, coef \%*\% theta = fpcs}
\item{convergence}{logical, algorithm converged or not}
\item{no.iter}{number of iterations needed for convergence}
}
\description{
Decomposes functional observations using adaptive functional principal component analysis. This method incorporates adaptive smoothness into the estimated decomposition by using a penalized likelihood framework
}
\examples{
sim_data <- simulate_adaptive_functional_data(n.tp = 100)
afpca_output <- fpca.adapt(data = sim_data, nbs = 10)
}
\references{
Paper
}
