% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca.adapt.R
\name{fpca.adapt}
\alias{fpca.adapt}
\title{Adpative Functional Principal Component Analysis}
\usage{
fpca.adapt(
  data,
  poly.degree = 2,
  knots = NA,
  nbs = 35,
  n.comp = 18,
  normalize.scores = TRUE,
  orthogonalize_scores = TRUE,
  seed.num = 4,
  orthogonalize_fpcs = TRUE,
  ntimes = 100,
  basis = "trunc.poly",
  pve = 0.99
)
}
\arguments{
\item{pve}{}
}
\value{

}
\description{
Adpative Functional Principal Component Analysis
}
\examples{


}
