% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_basis.R
\name{generate_basis}
\alias{generate_basis}
\title{Generate Spline Basis to estimate FPC}
\usage{
generate_basis(
  data,
  basis = "trunc.poly",
  poly.degree = 3,
  knots = NA,
  nbs = 40
)
}
\arguments{
\item{data}{input data}

\item{basis}{Type of spline basis. "os" for B-spline basis and "trunc.poly" for truncated polynomials (default)}

\item{poly.degree}{Degree of spline basis}

\item{knots}{passed to AdaptFitOS}

\item{nbs}{number of spline basis}
}
\value{

}
\description{
This is an internal function to generate a spline basis used to calculate the FPCs using asp2 from AdaptFitOS
}
\keyword{internal}
