% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likelihood.R
\name{get_likelihood}
\alias{get_likelihood}
\title{Title}
\usage{
get_likelihood(
  data,
  coef,
  Theta,
  c_mat,
  N.Unp.Basis,
  nbs,
  n.comp,
  lambda,
  sigma
)
}
\arguments{
\item{sigma}{}
}
\value{

}
\description{
Title
}
\keyword{internal}
