% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likelihood.R
\name{get_likelihood}
\alias{get_likelihood}
\title{Estimate likelihood of the data}
\usage{
get_likelihood(
  data,
  coef,
  Theta,
  c_mat,
  N.Unp.Basis,
  nbs,
  n.comp,
  lambda,
  sigma
)
}
\arguments{
\item{data}{input data}

\item{coef}{spline coefficients}

\item{Theta}{spline basis}

\item{c_mat}{matrix of scores}

\item{N.Unp.Basis}{number of unpenalized spline basis}

\item{nbs}{total number of spline basis}

\item{n.comp}{number of components being estimated}

\item{lambda}{penalty function}

\item{sigma}{residual variance}
}
\value{

}
\description{
Internal function to estimate the likelihood of the data, used to check algorithm converges
}
\keyword{internal}
