% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthonormal_step.R
\name{orthonormal_step}
\alias{orthonormal_step}
\title{Function to orthonormalize (or orthogonalize) scores}
\usage{
orthonormal_step(c_mat, orthogonalize_scores, normalize.scores)
}
\arguments{
\item{c_mat}{matrix of scores}

\item{orthogonalize_scores}{logical; wether to orthogonalize scores}

\item{normalize.scores}{logical; wether to normalize scores}
}
\value{

}
\description{
Internal function to orthonormalize or orthogonalize scores
}
\keyword{internal}
