% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_adaptive_functional_data.R
\name{simulate_adaptive_functional_data}
\alias{simulate_adaptive_functional_data}
\title{Simulate Adaptive Functional Data}
\usage{
simulate_adaptive_functional_data(
  n.tp = 200,
  N.subj = 20,
  var.scores = c(4, 1),
  noise.var = 0.2,
  seed.num = 1
)
}
\arguments{
\item{n.tp}{number of timepoints}

\item{N.subj}{number of subjects}

\item{var.scores}{variance of the scores (this is the value of the eigenvalues)}

\item{noise.var}{residual noise variance}

\item{seed.num}{seed number}
}
\value{
An object of class "fpca_sim_data" that contains the following:

\item{data}{simulated dataset}
\item{data_true}{simulated dataset without any noise}
\item{Mu_true}{The true data-generating mean-function}
\item{Phi_true}{The true data-generating functional principal components}
\item{Scores_true}{The true scores}
}
\description{
Simulated functional data that includes with sharp changes over the domain of the function (i.e. smoothness rapidly changes over the domain). The data simulates data using sinusoidal functions with varying period and amplitude.
}
\examples{
sim_data <- simulate_adaptive_functional_data(n.tp = 100, N.subj = 25)
}
